+++ src/mol-plugin/util/headless-screenshot.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2019-2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2019-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
  * @author Jesse Liang <jesse.liang@rcsb.org>
@@ -209,6 +209,7 @@ export function defaultImagePassParams(): Partial<ImageProps> {
             ...DefaultCanvas3DParams.multiSample,
             mode: 'on',
             sampleLevel: 4,
+            reuseOcclusion: false,
         }
     };
 }
@@ -221,6 +222,7 @@ export const STYLIZED_POSTPROCESSING: Partial<PostprocessingProps> = {
             radius: 5,
             bias: 0.8,
             blurKernelSize: 15,
+            blurBias: 0.5,
             resolutionScale: 1,
             color: ColorNames.black,
         }