+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Note that since we don't clearly distinguish between a public and private interf
 
 - Add depth of field (dof) postprocessing effect
 - Add `SbNcbrTunnels` extension for for visualizing tunnels in molecular structures from ChannelsDB (more info in [tunnels.md](./docs/docs/extensions/tunnels.md))
+- Fixes SSAO edge artifacts (#1122)
+    - Add `reuseOcclusion` parameter to multi-sample pass
+    - Add `blurBias` parameter to occlusion pass
 
 ## [v4.2.0] - 2023-04-05